% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_feature_stemmer.R
\name{ft_stemmer}
\alias{ft_stemmer}
\title{Feature Transformation -- Stemmer (Transformer)}
\usage{
ft_stemmer(x, input_col, output_col, language = "English",
  uid = sparklyr::random_string("stemmer_"), ...)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_col}{The name of the input column.}

\item{output_col}{The name of the output column.}

\item{language}{A character string used to specify the language of a stemmed text.}

\item{uid}{A character string used to uniquely identify the feature transformer.}

\item{...}{Optional arguments; currently unused.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns a \code{ml_transformer},
  a \code{ml_estimator}, or one of their subclasses. The object contains a pointer to
  a Spark \code{Transformer} or \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the transformer or estimator appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, a transformer is constructed then
  immediately applied to the input \code{tbl_spark}, returning a \code{tbl_spark}
}
}
\description{
A stemmer that converts input tokens to their word stem.
}
\examples{
\dontrun{
library(dplyr)
library(sparklyr)
sc <- spark_connect(master = "local")
data.frame(text = "Mary had a little lamb") \%>\%
copy_to(sc, ., "test") \%>\%
sdf_mutate(tokens = ft_tokenizer(text),
           stoppedTokens = ft_stop_words_remover(tokens),
           stemmedTokens = ft_stemmer(stoppedTokens))}

}
\seealso{
\url{https://github.com/master/spark-stemming}

\url{http://snowballstem.org/}
}
